create services/ folder. move logger and db to services
